attendance,reviews added/ tournaments deleted

+++ src/main/java/ru/cs/ifmo/ligos/security/oauth2/UserPrincipal.java
@@ -12,13 +12,13 @@ import java.util.List;
 import java.util.Map;
 
 public class UserPrincipal implements OAuth2User, UserDetails {
-	private Integer id;
+	private Long id;
 	private String email;
 	private String password;
 	private Collection<? extends GrantedAuthority> authorities;
 	private Map<String, Object> attributes;
 
-	public UserPrincipal(Integer id, String email, String password, Collection<? extends GrantedAuthority> authorities) {
+	public UserPrincipal(Long id, String email, String password, Collection<? extends GrantedAuthority> authorities) {
 		this.id = id;
 		this.email = email;
 		this.password = password;
@@ -43,7 +43,7 @@ public class UserPrincipal implements OAuth2User, UserDetails {
 		return userPrincipal;
 	}
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
